final version

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {FormGroup} from '@angular/forms';
 import {BasketService} from 'src/app/basket/basket.service';
 import {CheckoutService} from '../checkout.service';
@@ -7,35 +7,81 @@ import {IBasket} from "../../shared/models/basket";
 import {IOrder} from "../../shared/models/order";
 import {NavigationExtras, Router} from "@angular/router";
 
+declare var Stripe;
+
 @Component({
     selector: 'app-checkout-payment',
     templateUrl: './checkout-payment.component.html',
     styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     @Input() checkoutForm: FormGroup;
+    @ViewChild('cardNumber', {static: true}) cardNumberElement: ElementRef;
+    @ViewChild('cardExpiry', {static: true}) cardExpiryElement: ElementRef;
+    @ViewChild('cardCvc', {static: true}) cardCvcElement: ElementRef;
+    stripe: any;
+    cardNumber: any;
+    cardExpiry: any;
+    cardCvc: any;
+    cardErrors: any;
+    cardHandler = this.onChange.bind(this);
+    loading = false;
 
-    constructor(private basketService: BasketService, private checkoutService:
-        CheckoutService, private toastr: ToastrService, private router: Router) {
+    constructor(private basketService: BasketService, private checkoutService: CheckoutService,
+                private toastr: ToastrService, private router: Router) {
     }
 
-    ngOnInit(): void {
+    ngAfterViewInit(): void {
+        this.stripe = Stripe('pk_test_51KPzWSEi4HmFBiZK1KKhPIa8paOFiQtXia2nHq0BD8ORDlKjQ6wjPPa7Z3EBQu0J3xhpVWe7c2L5sK2PiRUNhmOQ00GExCrvSS');
+        const elements = this.stripe.elements();
+
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber.mount(this.cardNumberElement.nativeElement);
+        this.cardNumber.addEventListener('change', this.cardHandler);
+
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+        this.cardExpiry.addEventListener('change', this.cardHandler);
+
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc.mount(this.cardCvcElement.nativeElement);
+        this.cardCvc.addEventListener('change', this.cardHandler);
     }
 
-    submitOrder() {
+    ngOnDestroy() {
+        this.cardNumber.destroy();
+        this.cardExpiry.destroy();
+        this.cardCvc.destroy();
+    }
+
+    onChange({error}) {
+        if (error) {
+            this.cardErrors = error.message;
+        } else {
+            this.cardErrors = null;
+        }
+    }
+
+    async submitOrder() {
+        this.loading = true;
         const basket = this.basketService.getCurrentBasketValue();
-        const orderToCreate = this.getOrderToCreate(basket);
+        try {
+            const createdOrder = await this.createOrder(basket);
+            const paymentResult = await this.confirmPaymentWithStripe(basket);
 
-        this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
-            this.toastr.success('Order created successfully.');
-            this.basketService.deleteLocalBasket(basket.id);
-
-            const navigationExtras: NavigationExtras = { state: order };
-            this.router.navigate(['checkout/success'], navigationExtras);
-        }, error => {
-            this.toastr.error(error.message);
+            if (paymentResult.paymentIntent) {
+                this.basketService.deleteLocalBasket(basket.id);
+                const navigationExtras: NavigationExtras = {state: createdOrder};
+                this.router.navigate(['checkout/success'], navigationExtras);
+            } else {
+                this.toastr.error(paymentResult.error.message);
+            }
+            this.loading = false;
+        }
+        catch (error) {
             console.log(error);
-        });
+            this.loading = false;
+        }
     }
 
     private getOrderToCreate(basket: IBasket) {
@@ -44,5 +90,21 @@ export class CheckoutPaymentComponent implements OnInit {
             deliveryMethodId: +this.checkoutForm.get('deliveryForm').get('deliveryMethod').value,
             shipToAddress: this.checkoutForm.get('addressForm').value
         }
+    }
+
+    private async createOrder(basket: IBasket) {
+        const orderToCreate = this.getOrderToCreate(basket);
+        return this.checkoutService.createOrder(orderToCreate).toPromise();
+    }
+
+    private async confirmPaymentWithStripe(basket: IBasket) {
+        return this.stripe.confirmCardPayment(basket.clientSecret, {
+            payment_method: {
+                card: this.cardNumber,
+                billing_details: {
+                    name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+                }
+            }
+        });
     }
 }
